requirements.txt
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
config==0.4.2
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11
django-annoying==0.10.4
django-appconf==1.0.3
django-bootstrap3==11.0.0
django-bootstrap4==0.0.8
django-heroku==0.3.1
django-imagekit==4.0.2
django-location-field==2.0.6
django-registration==2.4.1
django-tinymce==2.7.0
gunicorn==19.9.0
hypothesis==4.23.0
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
olefile==0.46
pilkit==2.0
Pillow==6.0.0
psycopg2==2.8.2
psycopg2-binary==2.7.5
pycodestyle==2.5.0
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
whitenoise==4.1.2
wrapt==1.11.1
